Path updates
Now loading a number of paths from the environment. Using the path information to determine the position error of the vehicle relative to the active path.
Updated README to provide instruction on how to name/structure additional paths in the simulation environment
Updated simulation environment with "new" path naming/structure.