Simulation Environment, Steering Updates
Updated steering functions to have a maximum steering angle to go along with the change to the "Manta" vehicle in Coppelia sim.
Moved test script to a separate file. Implemented a simple steering controller for testing of the bridge.
Updated the Coppelia Simulation environment to the Manta vehicle, and added cameras to watch the vehicle.