Sim interface dev (#1)
* Updates and additions

Corrected object references to Coppelia sim objects.
Added a setSpeed and setSteering function for control of the sim

Add some quick test lines to experiment with the Coppelia sim interface

* Update readme, Added todo list, starting path info retrieval

Beginning of path info retrieval. More work to be done in parsing the raw data of points along the path.
Updated Readme to include basic instructions.

* Path updates

Now loading a number of paths from the environment. Using the path information to determine the position error of the vehicle relative to the active path.
Updated README to provide instruction on how to name/structure additional paths in the simulation environment
Updated simulation environment with "new" path naming/structure.

* Simulation Environment, Steering Updates

Updated steering functions to have a maximum steering angle to go along with the change to the "Manta" vehicle in Coppelia sim.
Moved test script to a separate file. Implemented a simple steering controller for testing of the bridge.
Updated the Coppelia Simulation environment to the Manta vehicle, and added cameras to watch the vehicle.

* Hide motorbike in Coppelia Sim

* Orientation Error Work

Orientation error is a WIP. A coordinate frame issue currently exists.
Updated Pathinformation.png to reflect current simulation structure.