Bridge, Qscene Upates
Updated Qscene3 to have a 0.05 S simulation time step. Stability seems good and this improves CPU time needed for an episode significantly.
Corrected vehicle orientation relative to the path in the InitPosition function
Added a sign to path error to inform the side of the path the vehicle is on.
Corrected vehicle position in path error to be in the path's frame rather than world frame.