RL agent Update (#11)
* Sim interface dev (#1)

* Updates and additions

Corrected object references to Coppelia sim objects.
Added a setSpeed and setSteering function for control of the sim

Add some quick test lines to experiment with the Coppelia sim interface

* Update readme, Added todo list, starting path info retrieval

Beginning of path info retrieval. More work to be done in parsing the raw data of points along the path.
Updated Readme to include basic instructions.

* Path updates

Now loading a number of paths from the environment. Using the path information to determine the position error of the vehicle relative to the active path.
Updated README to provide instruction on how to name/structure additional paths in the simulation environment
Updated simulation environment with "new" path naming/structure.

* Simulation Environment, Steering Updates

Updated steering functions to have a maximum steering angle to go along with the change to the "Manta" vehicle in Coppelia sim.
Moved test script to a separate file. Implemented a simple steering controller for testing of the bridge.
Updated the Coppelia Simulation environment to the Manta vehicle, and added cameras to watch the vehicle.

* Hide motorbike in Coppelia Sim

* Orientation Error Work

Orientation error is a WIP. A coordinate frame issue currently exists.
Updated Pathinformation.png to reflect current simulation structure.

* Rl utils (#2)

* initial testing

* Episode Handling

* TODO Tasks

* Initial Position and Orientation Error

Added function to CoppeliaBridge to set the vehicle's initial position to a point along the chosen path. Also corrected issues in orientation error. Now reports error about the Z axis.

* Added Collision Detection

Added a proximity sensor to the vehicle to enable collision detection.
Updated environment.py to use collision as an additional method for reset the episode.

* Simple RL Agent Framework

* Qcar added

* looks complete

* Initial Agent AI

* changes

* more fixes

* Draw io bridge diagrams (#6)

* Added drawIO diagrams to show Coppelia Bridge Structure and communication

* Updated TestScript.py to perform render time studies

Reformatted my test script to allow for the simulation to be run with and without rendering, record time for each run, and report the mean time. This was done to study the impact of rendering on the simulation time.

* Occupancy Grid is working

* Reward Function framework (#7)

Implemented framework for handling reward functions. This framework allows for choosing between multiple reward functions with only the change of a parameter. This minimizes the impact to the RL environment when a different reward function is attempted.

* attempted merge from main

* modified for RL Agent

* Start of merge fixes

Set init position is still broken. Needs to be looked in to.
Plotting of occupancy graph not working on my machine. Commented out plt.ion() for now.

* minor changes to QScenes

* Changed to Mujoco and Fixed setSteering function

* Single Lane better obstacles

* Reduced bounce

* Bridge, Qscene Upates

Updated Qscene3 to have a 0.05 S simulation time step. Stability seems good and this improves CPU time needed for an episode significantly.
Corrected vehicle orientation relative to the path in the InitPosition function
Added a sign to path error to inform the side of the path the vehicle is on.
Corrected vehicle position in path error to be in the path's frame rather than world frame.

* Added renderstate function, calling more inits in __init__

The renderState function sets the status of rendering. This is called in the startSimulation function with a new optional argument to set the render state. On by default.
Calling initEgo and initScene in __init__ to simplify the startup. These only need to be called once per bridge instance.

* Added path to the occupancy grid

Added pathForOG to generate positions in the occupancy grid which represent the path. Currently a straight line but could become a spline path in the future.

* Corrected Collision Detection

Collision detection is done by using a subset of the occupancy grid. Currently the vehicle dimensions have a bit of a buffer.
Changed logic in occupancy grid to have 0s represent free space and 1s represent obstacles.

* Updated TODOs

* Updates to Environment.py

Updated environment.py to match new bridge conventions
Fixed bounding of path on occupancy grid in the bridge.

* State handling NN architecture update

* Reward Calculator

* Updated Reward Calculator

* Cont. Actions and DDPG

* cleanup

* simple reward

---------

Co-authored-by: cdauchess <cdauche@clemson.edu>
Co-authored-by: cdauchess <charlie.dauchess@gmail.com>
Co-authored-by: pkorrapati <pranav.korrapati@gmail.com>
Co-authored-by: Pranav Korrapati <pkorrap@clemson.edu>