Further looking path information in the occupancy grid.
Added two methods to look further ahead along the path and display it on the occupancy grid. One represents the path as a series of bezier curves and one represents the path as a series of line segments. With our current 1.5M vision sensors the line segment version works better. There isn't significant curve in the path for the bezier curves to be effective.