Training  (#13)
* Trained Agent

300 Episodes trained agent with no initial position randomization.

* Added setInitPosition to the randomize step of reset vehicle in environment

This allows the environment to place the vehicle at a random point along the path to reset the vehicle.
Renamed check point file and plot image to 01 and added a training log.
Created QScene4 which has a smaller path that the vehicle can explore better

* Two additional training sessions complete.

02 and 03 agents trained. See training log for details.

* Reward and Episodes

* logs and path error

* Added QScene5 with a smaller path

Smaller path so that the vehicle covers more of it during training.

* Update QScene5.ttt

* Update Qscene5: Moved vehicle initial position to be on the path.

* New Occupancy

* Trained Agent 5 added.

This agent didn't train well. Some episodes were cut short due to starting on/near obstacles.

* Preventing starting on obstacles

Added a check to the initial position to prevent starting on obstacles. This offsets the start position if too near an obstacle.

* Updates

Updated configs with new reward scales, learning rates.
Turned on add noise to RL agent by default
Changed path error exponential to -5 from -2
Updated top camera view to see edges of path
Moved training metrics images to the trained agents folder

---------

Co-authored-by: Kalpit-Vadnerkar <kvadner@g.clemson.edu>